More cleaning and added docs.
We’re ready for review
