+++ index/circular_reference_result.go
@@ -14,6 +14,7 @@ type CircularReferenceResult struct {
 	IsInfiniteLoop      bool   // if all the definitions in the reference loop are marked as required, this is an infinite circular reference, thus is not allowed.
 }
 
+// GenerateJourneyPath generates a string representation of the journey taken to find the circular reference.
 func (c *CircularReferenceResult) GenerateJourneyPath() string {
 	buf := strings.Builder{}
 	for i, ref := range c.Journey {
